apps/api/requirements.txt
# FastAPI Backend Dependencies

# Core
fastapi==0.128.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic>=2.7.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# NLP (reference to packages/nlp)
# sentence-transformers and scikit-learn are in packages/nlp

# Utilities
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
ruff==0.1.13
